Making get_name_email() method more robust by accounting for invalid email addresses.
We're trying to get the name out the email address if we have nothing else to work with.  Unfortunately not all email addresses are in a valid format.  This causes the split method to only return one value sometimes as opposed to 2 breaking the call.  I've added an if statement that can catch one variable or two.  This prevents the code from breaking on an invalid email address.