Changing title of "Membership Joins Summary Report" to "Memberships by Join Date."
Updating the membership report form to include the new start and end date fields.

Removing the 3 sets of membership sections and replacing it with one big list of memberships that are filtered by the form fields.

Changing table field names to names that fit on one line.  Adding color coding dependent on the status of the membership record (e.g. active, expired, archived, etc …)