Updating membership application
These recent updates label whether the membership type your applying for is considered a "join" or "renewal."

We also do a more accurate job of displaying only the membership types that should be available to use as a superuser or anyone else.

Updated the logic that renews members via the membership application.