Rework dependencies and pip install
The motivation for this was to remove the overly complex (and incorrect) easy
install instructions and replace them with the modern way of doing pip
installs.

While at it I decided some of the segmentation didn't make sense (like calling
it 'pre config') so I renamed and rearranged a couple of related sections.