fix user merge error
The error occurred because the functions "get_all_related_objects" and "get_all_field_names" are deprecated and not returning the exact same items as they used to. Resolved by replacing those two functions with get_fields(). In addition, the function was not handling many-to-many fields. Fixed that as well.