Commented Out Share Link
The "Share" link shows up by default in staff under staff profile. The link itself doesn't go anywhere and there is no front end way to update link so that you can utilize it to send or share information. Commenting out instead of deleting as we may want to utilize function in the future for generating an email to share a staff member profile.