Make the tags list work with Django 1.8+
Django 1.8 removed the 'name' field on the ContentType model and
replaced it with a 'name' property.  This causes a FieldError
"Cannot resolve keyword 'name' into field. Choices are: app_label, ..."
when .order_by('name') is called, since .order_by() is translated into
SQL that can only operate on model fields (database columns).

To fix this, use sorted() to perform the sort in Python code after the
data is returned by the database.