Fix Preview when adding a new Page
If a new Page is previewed before it is actually added, preview() in
views.py calls page=form.save(commit=False), which populates app_object
in the top_nav_items.html template with a temporary page object that
has no primary key (because it has not been committed yet).
top_nav_items.html checks for an empty app_object before using it, but
does not check for an empty app_object.pk before using that, which
leads to an error.