Fix Bootstrap 3 .label CSS conflicts
A number of forms used by various Tendenci apps contain <label> tags
wrapped in the 'label' CSS class.  This conflicts with the Bootstrap 3
'label' class, which is intended to be used to for labels next to
headings (not for labels in forms).  This conflict results in unreadable
form labels on a number of Tenndenci pages.

It appears that Bootstrap 3 "Labels" are now called "Tags" in Bootstrap
4, so this problem should resolve itself when Tendenci upgrades to
Bootstrap 4.

However, as an interim solution, a simple fix is to modify Bootstrap
such that the Bootstrap 3 'label' class will not be applied to <label>
tags.