Fix events search_indexes failure due to django_countries iteration
The wsgi loader calls django.setup(), which loads the haystack app,
which calls QueuedSignalProcessor.setup() in
tendenci/apps/search/signals.py, which imports search_indexes.py in the
events Tendenci app, which imports forms.py, which imports
tendenci.apps.base.forms, which iterates through
django_countries.countries, which causes the django_countries module to
attempt to translate the country names.  This fails because translations
are not available until after django.setup() has completed.

This breaks the search index for the events Tendenci app.

The events forms are not actually needed by search_indexes.py, so simply
ignore the exception that is thrown when they are imported during Django
initialization.