Fix handling of hidden folders in theme_editor
The previous code would throw an exception if a non-hidden folder was
nested inside a hidden folder.  This is because os.walk() traverses into
hidden folders, and the code searches for the parent of each non-hidden
folder in files_folders, but it doesn't add hidden folders to
files_folders, and it doesn't handle the case where the parent is
not found in files_folders.  So a non-hidden folder in a hidden folder
causes the parent search to fail.