Fix view permissions on objects with 1 digit primary keys
If search indexes are available, has_perm() in apps/perms/backend.py
calls can_view() in apps/perms/utils.py to determine whether a user has
view permissions on an object.  can_view() queries the index for
documents matching (SQ(primary_key=obj.pk) & (q_groups | q_users)) and
returns True if any matching documents are found.

Since the primary_key index field is defined as a CharField, Whoosh
applies StopFilter to the query for that field.  If the query value is
less than 2 characters long (if the object's primary key happens to be
<= 9) then StopFilter removes that query from the query set, reducing
the query to just (q_groups | q_users).

That causes documents to be returned (and can_view() to return True) if
the user has view permissions on any objects in the index.

This commit fixes the issue by changing primary_key to an IntegerField
such that StopFilter will not be applied to queries on that field.

In addition, a sanity check is added to can_view() to help ensure that
the result set returned by the index makes sense.