Use new formatting for checkbox fields on membership forms
The formatting of individual checkbox fields is inconsistent with other
fields on membership forms.  It appears that individual checkbox fields
use old formatting, while all other fields have been updated to use new
formatting.  It also appears that the new formatting was intendend to
apply to individual checkboxes, but an old code path caused the old
formatting to continue to be used.

This commit disables that old code path so that individual checkboxes
are formatted more like other fields.