Replace deprecated request.REQUEST
request.REQUEST is removed in Django 1.9 and does not work when running
Django 1.8 under the Daphne web server.  This commit replaces its use
with request.POST/request.GET as appropriate, and also switches from the
abandoned django-pagination library to the dj-pagination fork to fix the
use of request.REQUEST in that library.  (May want to consider switching
from dj-pagination to the newer Pagination feature in Django itself.)