Allow user to view/edit invoice if bill_to_email matches
If a user registers for an event without logging in, an invoice is
generated with no creator or owner.  If the user then logs in, "My
Invoices" will list the invoice because bill_to_email matches the user's
email.  However, if the user clicks on the invoice, they get the
"Forbidden" page because the permissions checks only look at
creator/owner and not bill_to_email.

This commit updates the permissions checks to allow the user to
view/edit the invoice if bill_to_email matches the user's email.