SECURITY: Disable GZipMiddleware to prevent BREACH attacks
"BREACH" refers to a class of attacks that use on-demand gzip encoding
to determine the content of an encrypted connection without actually
decrypting it.  Due to this class of attacks, it is not safe to use gzip
encoding for most encrypted dynamic content, and therefore it is not
safe to use GZipMiddleware to unconditionally gzip encode all content
generated by Django.

See http://breachattack.com/ and
https://docs.djangoproject.com/en/1.11/ref/middleware/#module-django.middleware.gzip
for more info.