Fix issues with "of current year" fixed expiration dates
If the specified month/day for an "of current year" fixed expiration
date has already passed and a grace period is not configured, the
previous logic would cause the membership to be expired when it was
created.  Similarly, if the specified month/day has already passed and a
grace period is configured that crosses a year boundary, the previous
logic would not handle the grace period properly at the end of the year.
This is fixed by adding a year to the expiration date if the specified
month/day has already passed.

In addition, the grace period calculation had an off-by-one error
related to the rounding of times to days.  This is fixed by using <
instead of <= when checking the grace period.