Move boilerplate top_nav HTML to the common template
The top_nav.html templates all contain similar HTML for rendering the
nav button, but there are a large number of inconsistencies and errors.

Moving this boilerplate HTML to a single template will eliminate the
inconsistencies and make it easier to update the HTML, for example to
work with new versions of bootstrap.