Defer executing DB queries
When importing memberships.reports or model_report.utils the database queries
are run immediately.
This causes problems when run through pytest-django (and possibly other tools)
which suppress unrequested DB access as a matter of course

I've opted to defer the queries here as I saw the same method used elseware in
those files and thought it was likely to be an acceptable solution.