Disable code that uses abandoned Geraldo dependency
Geraldo is abandoned and does not support Python 3.  reportlab can
technically implement the same functionality under Python 3, but its
API for handling this use case is much more complex than Geraldo, so
converting would involve significant effort.

The only place Geraldo is currently used is in the "Memberships by
Join Date" report, to generate a downloadable PDF version of the
report.  However, the generated PDF is ugly and unreadable (text in
adjacent columns within a single row overlap each other).  All of the
information on this report is also available via the "Active
Memberships Report", which supports a downloadable CSV version.

For now, just disable the PDF download button on the "Memberships by
Join Date" report.