Update deprecated render_to_response() calls
The context_instance parameter is deprecated as of Django 1.8.  Context
should now be provided as a dict using the context parameter.  If the
template requires the request, then render() should be used in place of
render_to_response().  render_to_response() is also deprecated as of
Django 2.0.

One of the nice things about render_to_response() was that it was easy
to distinguish between it and other unrelated but similar looking
render() methods.  For example, the render() method on objects returned
by django.template.loader.get_template() also takes context and response
parameters, but does not take the other parameters that
django.shortcuts.render() takes, which can be confusing if you confuse
the two render() methods.  Similarly, objects returned by
engines['django'].from_string() also have a render() method, but it
takes a single Context or RequestContext parameter.

To minimize confusion, this commit uses
`from django.shortcuts import render as render_to_resp` to rename the
new method to a more unique name.  Note that render() and
render_to_response() take different parameters, so simply renaming
render() to render_to_response() might confuse anyone who is still
expecting the old parameters to work.