Overhaul theme handling
themed_response() and render_to_theme() need to be updated to use the
new django.shortcuts.render() parameters.

However, the existing implementation looks like it hasn't worked
as designed at least since Tendenci was updated to Django 1.8.  There
is lots of dead code, it performs redundant template lookups, and there
are a number of unhandled corner cases.

This commit cleans up this code, updates it to work with modern Django,
and restores the ability to dynamically disable the theme using a
?disable_theme query parameter.