Django 1.11: Fix RemovedInDjango20Warnings about Manager inheritance changes
Classes inheriting from tendenci.apps.files.models.File did not
previously inherit the FileManager, although it appears that the code
assumed it would be inherited.

Django 2.0 changes the inheritance behavior of Managers on Models such
that those classes will inherit FileManager, and Django 1.11 prints a
warning about this change.

This commit opts-in to the new behavior to silence the warning and fix
any bugs related to the previous lack of inheritance.