Rename the whoosh index directory
Since the index needs to be rebuilt anyway for the Python 3 and Django
1.11 upgrades, this is a good time to rename it.

Whoosh wants to be able to delete and re-create the index directory,
which works better if the index is in a sub-directory.

In addition, index.whoosh appears very similar to index.html and may be
confusing to new users, so give it a name that is less confusing.