Add support for "builtin" themes
Currently, the Tendenci UI is split between the core UI, which is stored
in the Tendenci apps and is updated automatically with Tendenci, and
themes, which are typically downloaded once at installation time and
cannot be automatically updated once they have been customized.

Due to this split, it is not currently possible to make sweeping changes
to the core UI, such as upgrading Bootstrap, without breaking existing
themes.

To work around this, we could create a new theme which overrides much of
the core UI.  However, since themes cannot be automatically updated,
they should ideally override as little of the core UI as possible so
that users can benefit from updates and bug fixes in the core UI without
needing to manually update their themes.

A better solution would be to add support for multiple core UIs, which
can be selected based on the current theme.

To take this a step further, if we think of these core UIs themselves as
themes, we could add support for generic theme inheritance/layering, and
support for built-in themes that are automatically updated with
Tendenci, which would have numerous benefits.  For example, the
standard/example themes could be updated with Tendenci, user themes
could be reduced to a minimal set of overrides/customizations to the
standard themes, users would have easy access to the latest versions of
the original unmodified themes in the theme editor for comparison with
their own themes, and users could develop and share themes that could be
further overridden/customized by other users without affecting those
other users' ability to install updates or bug fixes provided by the
theme's developer.

This commit is the first step down that path, adding basic support for
built-in themes that will be automatically updated with Tendenci.