Add support for implicit static file overrides in themes
Currently, static files in Tendenci apps are referenced from templates
using {% static %} or {{ STATIC_URL }} or {{ LOCAL_STATIC_URL }}, and
static files in themes are referenced from templates using
{{ THEME_URL }}.  In order to override static files in a theme, all
templates which reference those static files must be copied to the theme
and updated to use the theme's static files.

This makes it difficult to override static files in a theme, and it
doesn't work well when theme inheritance is used.

This commit overrides the standard {% static %} tag to behave like the
theme template_loader and search for static files in themes before using
the copies in apps.  This also adds a builtin {% local_static %} tag to
return local URLs when used with a CDN.

Note that {{ STATIC_URL }}, {{ LOCAL_STATIC_URL }}, and {{ THEME_URL }}
cannot be updated to support this searching behavior, so they are now
deprecated and should not be used in any apps or themes going forward.
Similarly, {% get_static_prefix %} cannot search determine the correct
static prefix since it doesn't know what file to search for, so it
should be avoided as well.