avoid saving as null for the string-based fields in donations
If both blank=True and null=True, Django will store empty values as NULL in the database. This is causing error in donation_inv_add where  invoice.ship_to_city is not null.
Preferably, null=True should be removed from the string-based fields, but this cannot be done in one migration without causing the database error if there is a null value in the column to be altered.