Add yes option to pip uninstall in documentation (#697)
* Pre confirm removal of tendenci modules

There is no reason to check the user wants to remove these via pip; they've
either decided to by copying the command in to their shell or are going to do
it anyway because thats what the docs say.

* Use -y to pip uninstall for tendenci7 upgrade

My previous commit in this branch adds this to the T7.1 upgrade docs but missed
this file. The same reasoning applies - people have copied the commands from
these documents; they've already decided they want to do this.