Document testing requirements and usage
This tries to include the information found in #703, #704 and the documentation
component for #593 .

There are a lot of tests for helpdesk which fail with a ProgrammingError as the
migrations are not run by default. The passage added in this commit recommends
enabling helpdesk - and for completeness of testing that is probably a good
thing.

This is an omnibus of three commits ( 4b81ee773 ec53a2b67 d2d3ab4f4 ) in the
Tendenci 7 branch which have been combined for submission to Tendenci-next.