fix boolean settings
Although the data type is boolean, the value and default_value should be one of the values in the "true,false" (which is a string) to make it stick in the true/false dropdown. Otherwise, it would create some undesired effect. For example, the "Users Can Add Directories" is a boolean setting that defaults to false. Without the quotes around the false, it is loaded as False in the database and would appear as "true" from the settings page /settings/module/directories/#id_usercanadd because False is not an option in the select dropdown.