Moving meta footer
Moved meta to have it visible in the events module when logged in and without the locations needing to be there to view it. Added "t-footer-meta" class to better target that area for styles if need be. I wonder if the 'groups' was within the /events/meta.html file in legacy versions.