Allow Pages and Stories to appear in apps menu
Currently the Pages and Stories apps are restricted in the Apps navigation menu to
those with superuser access but its not actually required. All users are able
to search - indeed the Pages app displays in the top bar anyway - and should be
able to add pages with appropriate permissions in place.

This change removes the navigational restriction on non superusers.