fix InvalidCacheKey error
The error occurred because the max_length of the file field is set way too high (260) in the File model, while Memcached does not accept keys longer than 250. This fix resolves the error and  also limits the length of the file name to be less than or equal to 100 on file upload.