fix x_type initial
Moved the initial assignment of x_type to prepare_firstdatae4_form. This is because PaymentForm class is loaded only when the site is up and the value of x_type is based on the site setting, so it is not a good place to set the initial for x_type.