Added support of a "groups" field on Membership imports
Removed some bothersome whitespace that crept in somehow.

Respect "groups" when specified explicitly in import

On second thoughts prefer "groups" to be respected regardless of whether the membership is active. That is a decision we make when preparing the import file, if we want groups we specify them, if not we don't. We don't need an overriding activity check.