Fixed the confusion about user_groups
Turns out group_set didn't work, probably because of an ambiguity, namely the Group model has three whole foreign references to User and none hd related_names. So added a related_name for the user_groups and suddenly everything worked. 

Implemented reversible sorting.

Could not find a way to do that with a generic user_group list, so only in the postgresql case is the user_group column orderable.