Size  member add select boxes more reasonably (#951)
* Size  member add select boxes more reasonably

Alas these are left to the default select sizing and in CSS what we see happening is Django applying a height (which seems a reasonable height) and then bootstrap overriding that to auto and the size is browser default. 

There is possibly a "right" way to do this with CSS and Bootstrap, but I haven't got all day to research it. This is by any measure a LOT better than the default sizing I'm seeing.