Support searching for members NOT in any groups (#946)
* Support searching for members NOT in any groups

Improved the quality of the group filter queries too. Discovered the cause of the group_set failure, in that there was ambiguity in it. Creating an explicit related_name, fixed that.

* Replaced group_set with user_groups

* group_set -> user_groups refactoring

* More group_set to user_groups

* update the migration file with the related_name

The related_name change does not affect the database, but django will prompt for a new migration anyway. This update avoids that reflex.

Co-authored-by: Jenny Qian <jqian@tendenci.com>