Added some very useful feedback on parsing to import preview (#952)
* Added some very useful feedback on parsing to import preview

Specifically membership_type (provided as a key, is converted if possible to provide feedbck) and join_dt and expire_dt, are parsed as they will me on import to provide visual preview on successful parse.

* Fixed a couple of bugs

'membership_type' in data_list is a bit funny in that it seems to an int where everything else, including membership_number is  and the dates are strings. Took a  sledge hammer approach to testing it that will survive it's being an int or str.

* More specific exception handling

Fair enough, LGTM.com don't like generic exception handling. No problem, it's better rigour to consider possible failure modes indeed.

* Reduced database hits on membership_type preview

* Bug fix