create a workaround for newsletter without fullpage plugin
The fullpage plugin is removed since tinymce version 6, which means the html header, along with the style block will be removed in tinymce editor. This workaround allows style tag on newsletter edit, then moves the style block to html header on newsletter sending. In addition, it preserves the style for the newsletters created prior to tinymce version 6.