add option to allow event select multiple groups with one primary group
If multiple groups are selected, the primary group is used for financial reports.
Changed the groups field (ManyToManyField) to use a through model so that we can add an additional field is_primary.
Migration uses SeparateDatabaseAndState to rename the existing relationship table events_event_groups to events_eventgroup to avoid losing the existing relations.