USER_IS_STAFF for reports seems more reasonable than USER_IS_SUPERUSER
In the ideal this would be configurable. But too many things in tendenci are only available to website admins (super users) and association reports clearly should be available to committee members and/or staff etc. At least in any Association I've worked it. They don't all need or want to be website super users/admins.