resolve migration warnings
Resolved migration warnings regarding the fields creator, owner, entity, etc. The "makemigrations" creates a lot of files with changes for these fields, but these changes do not actually affect the database. So, we manually picked the changes from these files and added to their corresponding initial migrations to avoid those many files being (unnecessarily) added to the repo.

Below is the warning message when run the "migrate" command:
  Your models in app(s): 'articles', 'announcements', 'boxes', 'careers', 'case_studies', 'categories', 'chapters', 'committees', 'contacts', 'contributions', 'corporate_memberships', 'directories', 'discounts', 'educations', 'email_blocks', 'emails', 'events', 'files', 'forms', 'forums', 'help_files', 'helpdesk', 'industries', 'jobs', 'locations', 'memberships', 'navs', 'news', 'pages', 'photos', 'products', 'projects', 'regions', 'reports', 'resumes', 'speakers', 'staff', 'stories', 'stripe', 'studygroups', 'testimonials', 'trainings', 'user_groups', 'videos' have changes that are not yet reflected in a migration, and so won't be applied.
  Run 'manage.py makemigrations' to make new migrations, and then re-run 'manage.py migrate' to apply them.